+++ backend/Kamp/Controllers/KampTarifeYonetimController.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using STYS.Kamp.Dto;
 using STYS.Kamp.Services;
+using STYS.Licensing;
 using TOD.Platform.AspNetCore.Authorization;
 using TOD.Platform.AspNetCore.Controllers;
 using TOD.Platform.Licensing.AspNetCore;
@@ -10,7 +11,7 @@ namespace STYS.Kamp.Controllers;
 
 [Route("ui/kamptarife")]
 [ApiController]
-[RequiresLicensedModule("Kamp")]
+[RequiresLicensedModule(StysLicensedModules.Kamp)]
 public class KampTarifeYonetimController : UIController
 {
     private readonly IKampTarifeYonetimService _service;